agentsgym-py/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "agentsgym-py"
version = "0.1.0"
description = "Python client for the agentsgym skills registry service."
readme = "../README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Christoph Biering" }]
keywords = ["agents", "skills", "registry", "client"]
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
  "mypy>=1.10",
  "ruff>=0.6",
]

[project.urls]
Repository = "https://github.com/biering/agentsgym"

[tool.hatch.build.targets.wheel]
packages = ["agentsgym_py"]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
  ".venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true
files = [
  "agentsgym_py",
]